refactor: testing chat socket logic
